fix filtering of changeset information
After a transaction was applied, the resulting changeset was filtered
by the websocket handler that the transaction was submitted to instead
of the handler that the respective changeset was sent out with.

This resulted in new waypoints and annotations appearing for both
the red and the blue coalition.

This fixes one cause of https://github.com/jboecker/dcs-mission-planner/issues/23.